part 1 : read config from file. Next step, get resource map

+++ src/main/java/com/example/ComponentState.java
@@ -4,40 +4,41 @@ package com.example;
  * Created by wiz on 6/1/18.
  */
 public class ComponentState {
-    int in;
-    int out;
-    int allocated;
-    int cpuUsed;
+    long in;
+    long out;
+    long allocated;
+    long cpuUsed;
 
-    public int getIn() {
+    // getter and setter
+    public long getIn() {
         return in;
     }
 
-    public void setIn(int in) {
+    public void setIn(long in) {
         this.in = in;
     }
 
-    public int getOut() {
+    public long getOut() {
         return out;
     }
 
-    public void setOut(int out) {
+    public void setOut(long out) {
         this.out = out;
     }
 
-    public int getAllocated() {
+    public long getAllocated() {
         return allocated;
     }
 
-    public void setAllocated(int allocated) {
+    public void setAllocated(long allocated) {
         this.allocated = allocated;
     }
 
-    public int getCpuUsed() {
+    public long getCpuUsed() {
         return cpuUsed;
     }
 
-    public void setCpuUsed(int cpuUsed) {
+    public void setCpuUsed(long cpuUsed) {
         this.cpuUsed = cpuUsed;
     }
